pyproject.toml
[tool.poetry]
name = "pyscora-wrangler"
version = "1.1.5"
description = "Python lib for DE"
authors = ["Oncase <suporte@oncase.com.br>"]
maintainers = ["Guilherme Morone <guilherme.morone@oncase.com.br>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/oncase/pyscora-wrangler"
repository = "https://github.com/oncase/pyscora-wrangler"
keywords = ["wrapper", "scora", "python", "data_engineering"]
classifiers = [
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
include = ["LICENSE"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
boto3 = "^1.26.108"
awswrangler = "^2.20.1"
asyncio = "^3.4.3"
pyathena = "^2.23.0"
pyyaml = "^6.0"
ldap3 = "^2.9.1"
typeguard = "^3.0.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[tool.black]
line-length = 120
skip-string-normalization = true

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"
